api/requirements.txt
aniso8601==9.0.1
blinker==1.6.2
cachelib==0.9.0
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Flask==2.3.3
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.7.1
Werkzeug==2.3.7
